Adapt Puma config to support Puma 7
The Puma configuration template needs to be adjusted to work with Puma
7:

1. The callback hooks have been renamed in
https://github.com/puma/puma/pull/3438:

    - `on_worker_boot` -> `before_worker_boot`
    - `on_worker_shutdown` -> `before_worker_shutdown`
    - `on_restart` -> `before_restart`

This commit adds a conditional block based on the Puma version to use
the new hooks.

2. With https://github.com/puma/puma/pull/3297, Puma 7 raises an
exception if Puma's options are accessed before they are fully
loaded. Currently we pass these options during loading of the Puma
config file, which then triggers this exception. These options are used
to determine whether Puma is in clustered mode by checking the number of
workers.

This commit uses a new method `set_puma_worker_count` added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/204104.

Note that Puma 7.0.2 or higher is required because
https://github.com/puma/puma/pull/3731 fixes an upstream bug to make
Puma 7 work with `queue_requests` set to `false`.

`set_puma_worker_count` was added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/204104.

The GDK config was also adapted similarly in
https://gitlab.com/gitlab-org/gitlab-development-kit/-/merge_requests/5249.

Changelog: changed