Add nginx['default_server_enabled'] configuration parameter
In https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/8805 we
made the GitLab server the default HTTP server because many users were
relying on that behavior, which only worked by accident because of the
naming of configuration files. However, some users have chosen to define
a catch-all in `nginx['custom_nginx_config']` to disable all requests
that do not use the explicit hostname.

To support this use case, this commit adds a
`nginx['default_server_enabled']` that can be set to disable the
behavior in
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/8805.

Changelog: added