Use chpst to force regstry user on migrations
This will facilitate use cases when the user is running as root and
calls the registry database commands. This was troublesome in cases
where PostgreSQL was running on the same node as the registry and doing
peer authentication. Since the peer authentication expects the registry
system user instead of root, it would fail the connection or required
the user to run the migration command with `sudo -u registry` or `su -
registry`.

With this change, the users can always call `gitlab-ctl
registry-database migrate` commands without having to prefix it with a
user change. This should work for any case.

Changelog: added