Add support for registry database prefer mode
The registry database will implement a prefer mode in which it, if
registry.database.enabled is set to prefer, it will decide on
based on preexisting lockfiles whether it should use the filesystem
metadata or the database.

If the filesystem lockfile exists, it will use the filesystem.

If the database lockfile exsist, it will use the database.

If no lockfile exists, it will use the database.

Although, when the lockfile does not exist, it means the user has never
used the registry before. In this situation, there's a chance that the
user does not have a registry database configured yet. We only
autocreate the database if the user is using the GitLab-managed
PostgreSQL database. On external PostgreSQL, we can't guarantee that the
user has manually create the database, so we can't set the registry to
the prefer mode, otherwise the migrations would fail to execute.

The solution is only allow the `prefer` mode if we know that the user
has enabled the GitLab-managed embedded PosgreSQL.

Migrations WILL run:
- database.enabled = true with embedded PostgreSQL
- database.enabled = true with external PostgreSQL
- database.enabled = 'prefer' with embedded PostgreSQL - New prefer mode

Migrations WON'T run:
- database.enabled = false - Database disabled explicitly on that node
- database.enabled = 'prefer' with external PostgreSQL - Can't safely
run migrations without GitLab-managed DB. Also on Multi-node, we can't
be sure that the GitLab-managed DB was deployed as the registry might be
running on a separate node then PostgreSQL, so the migrations won't
automatically run unless explicitly configured by the user.

Changelog: added