[CI] Fix the builder image tags for the check-packages jobs
When container images are used from the BUILDER_IMAGE_REGISTRY the tag
used is always BUILDER_IMAGE_REVISION. Except for the check-packages
jobs where the latest tag is used. This switches the latest tag to the
BUILDER_IMAGE_REVISION variable tag to ensure building and testing uses
the same container image.