Fail reconfigure if SELinux commands cannot be run
Previously if SELinux commands did not succeed, the `gitlab-ctl
reconfigure` step continued without flagging an issue. This made it
appear as though RHEL 10 worked fine, but GitLab did not properly
configure the SELinux policy. This occurred because RHEL 10 was not
actually supported in the `SELinuxDistroHelper.selinux_supported?`
check.

Add a `set -e` to the bash script to fail if the SELinux policy
cannot be configured.

Relates to
https://gitlab.com/gitlab-org/omnibus-gitlab/-/work_items/9471

Changelog: fixed